🚚 api: first route : move getCertificationReports
method to src course controller directory
